Made focus requests of components a bit smarter (see below for details)
Imagine the scenario that the app thinks already of any word and the
'Apply' button is enabled. If you hit Enter (and don't change the
current sequence), the current guess will be considered as wrong and the
next guess will be computed.

If you hit the space bar, the current guess will be applied and the next
guess is computed.

If you hit any other key, the focus is transferred to the current
sequence